Add support for specifying a directory ID to Connect-FederatedDirectory.ps1
This commit adds a new parameter, `$DirectoryID`, to the `Connect-FederatedDirectory.ps1` script, allowing the user to specify a directory ID when connecting to the federated directory. The directory ID is included in the request body sent to the OAuth2 token endpoint. The commit also includes a minor fix to remove empty values from the request payload before sending the request.